+++ src/DatabaseProvider/index.d.ts
@@ -2,6 +2,8 @@ declare module '@nozbe/watermelondb/DatabaseProvider' {
   import * as React from 'react'
   import Database from '@nozbe/watermelondb/Database'
 
+  export const DatabaseContext: React.Context<Database>
+
   export interface DatabaseProviderProps {
     children?: React.ReactChild // only one child is allowed, goes through React.Children.only
     database: Database