+++ src/decorators/date/index.d.ts
@@ -1,7 +1,7 @@
 declare module '@nozbe/watermelondb/decorators/date' {
-  import { Decorator, RawDecorator } from "@nozbe/watermelondb/utils/common/makeDecorator";
-  import { ColumnName } from "@nozbe/watermelondb";
+  import { Decorator, RawDecorator } from '@nozbe/watermelondb/utils/common/makeDecorator'
+  import { ColumnName } from '@nozbe/watermelondb'
 
-  const date: Decorator<[ColumnName], (columnName: ColumnName) => RawDecorator>;
-  export default date;
+  const date: Decorator<[ColumnName], (columnName: ColumnName) => RawDecorator>
+  export default date
 }