+++ src/decorators/immutableRelation/index.d.ts
@@ -1,14 +1,11 @@
 declare module '@nozbe/watermelondb/decorators/immutableRelation' {
-  import { ColumnName, TableName } from "@nozbe/watermelondb";
-  import { Decorator, RawDecorator } from "@nozbe/watermelondb/utils/common/makeDecorator";
+  import { ColumnName, TableName } from '@nozbe/watermelondb'
+  import { Decorator, RawDecorator } from '@nozbe/watermelondb/utils/common/makeDecorator'
 
-  const immutableRelation: Decorator<[
-    TableName<any>,
-    ColumnName
-    ], (
-    relationTable: TableName<any>,
-    relationIdColumn: ColumnName,
-  ) => RawDecorator>;
+  const immutableRelation: Decorator<
+    [TableName<any>, ColumnName],
+    (relationTable: TableName<any>, relationIdColumn: ColumnName) => RawDecorator
+  >
 
-  export default immutableRelation;
+  export default immutableRelation
 }