+++ src/decorators/readonly/index.d.ts
@@ -1,7 +1,7 @@
 declare module '@nozbe/watermelondb/decorators/readonly' {
-  import { Decorator, RawDecorator } from "@nozbe/watermelondb/utils/common/makeDecorator";
+  import { Decorator, RawDecorator } from '@nozbe/watermelondb/utils/common/makeDecorator'
 
-  const readonly: Decorator<[], () => RawDecorator>;
+  const readonly: Decorator<[], () => RawDecorator>
 
-  export default readonly;
+  export default readonly
 }