+++ src/decorators/relation/index.d.ts
@@ -1,17 +1,16 @@
 declare module '@nozbe/watermelondb/decorators/relation' {
-  import { ColumnName, TableName } from "@nozbe/watermelondb";
-  import { Decorator, RawDecorator } from "@nozbe/watermelondb/utils/common/makeDecorator";
-  import { Options } from "@nozbe/watermelondb/Relation";
+  import { ColumnName, TableName } from '@nozbe/watermelondb'
+  import { Decorator, RawDecorator } from '@nozbe/watermelondb/utils/common/makeDecorator'
+  import { Options } from '@nozbe/watermelondb/Relation'
 
-  const relation: Decorator<[
-    TableName<any>,
-    ColumnName,
-    Options | void
-  ], (
-    relationTable: TableName<any>,
-    relationIdColumn: ColumnName,
-    options: Options | void,
-  ) => RawDecorator>;
+  const relation: Decorator<
+    [TableName<any>, ColumnName, Options | void],
+    (
+      relationTable: TableName<any>,
+      relationIdColumn: ColumnName,
+      options: Options | void,
+    ) => RawDecorator
+  >
 
-  export default relation;
+  export default relation
 }