+++ src/hooks/test.js
@@ -6,6 +6,12 @@ import DatabaseProvider from '../DatabaseProvider'
 import Database from '../Database'
 import { MockProject, MockTask, MockComment } from '../__tests__/testModels'
 
+/**
+ * Note: this uses two testing libraries; react-test-renderer and react-hooks-testing-library.
+ * This is probably overkill for such a simple hook but I will leave these here in case more
+ * hooks are added in the future.
+ * */
+
 describe('useDatabase hook', () => {
   let database
   beforeAll(() => {