+++ src/observation/observeQuery.d.ts
@@ -1,8 +1,8 @@
 declare module '@nozbe/watermelondb/observation/observeQuery' {
-  import { Model, Query } from "@nozbe/watermelondb";
-  import { Observable } from "rxjs";
+  import { Model, Query } from '@nozbe/watermelondb'
+  import { Observable } from 'rxjs'
 
   export default function observeQuery<Record extends Model>(
     query: Query<Record>,
-  ): Observable<Record[]>;
+  ): Observable<Record[]>
 }