+++ CHANGELOG.md
@@ -15,8 +15,14 @@ All notable changes to this project will be documented in this file.
 - Updated demo dependencies
 - LokiJS is now a dependency of WatermelonDB (although it's only required for use on the web)
 - [Android] fix compile error for kotlin 1.3+ 
-- [hooks] new `useDatabase` hook for consuming the Database Context
-- [TypeScript] new type declarations
+- [Hooks] new `useDatabase` hook for consuming the Database Context:
+   ```
+   import { useDatabase } from '@nozbe/watermelondb/hooks';
+   const Component = () => {
+      const database = useDatabase();
+   }
+   ```
+- [TypeScript] added `.d.ts` files. Please note: TypeScript definitions are currently incomplete and should be used as a guide only. PRs for improvements would be appreciated.
 
 ## 0.10.1 - 2019-02-12